refactor(interface): add folders structure
